+++ extensions/dub/src/search-links.tsx
@@ -1,4 +1,3 @@
-import "@/polyfills/fetch";
 import { Action, ActionPanel, Alert, Color, confirmAlert, Icon, Keyboard, List, showToast, Toast } from "@raycast/api";
 import { ShortLinksResponse, useShortLinks } from "@hooks/use-short-links";
 import { DUB_CO_URL } from "@utils/constants";